Convert Default renderer into formatAuto service. Delete EuiValueRenderers module.
