--- src/components/value_renderer/compound/index.js
@@ -1,2 +1,0 @@
-export { property } from './property';
-export { join } from './join';
