--- src/components/value_renderer/compound/join.test.js
@@ -1,18 +1,0 @@
-import { join } from './join';
-
-describe('Value Renderer', () => {
-  describe('join', () => {
-
-    const array = [ 'a', 'b', 'c' ];
-    const render = (value) => value;
-
-    test('with default delimiter', () => {
-      expect(join(render)(array)).toBe('a, b, c');
-    });
-
-    test('with custom delimiter', () => {
-      expect(join(render, ' | ')(array)).toBe('a | b | c');
-    });
-
-  });
-});
