--- src/components/value_renderer/compound/property.js
@@ -1,5 +1,0 @@
-import { get } from 'lodash';
-
-export const property = (prop, renderer) => {
-  return (value) => renderer(get(value, prop));
-};
