--- src/components/value_renderer/default_renderer/index.js
@@ -1,1 +1,0 @@
-export { defaultRenderer } from './default_renderer';
