--- src/components/value_renderer/index.js
@@ -1,1 +1,0 @@
-export { EuiValueRenderers } from './value_renderers';
