--- src/components/value_renderer/value_renderers.js
@@ -1,8 +1,0 @@
-import { property, join } from './compound';
-import { defaultRenderer } from './default_renderer';
-
-export const EuiValueRenderers = {
-  default: defaultRenderer,
-  property,
-  join
-};
