+++ src/services/format/index.js
@@ -1,3 +1,4 @@
+export { formatAuto } from './format_auto';
 export { formatBoolean } from './format_boolean';
 export { formatDate } from './format_date';
 export { formatNumber } from './format_number';
